chore: rename samples project
