+++ .run/Samples.run.xml
@@ -1,13 +1,13 @@
 ﻿<component name="ProjectRunConfigurationManager">
-  <configuration default="false" name="OneWay (Sample)" type="DotNetProject" factoryName=".NET Project">
-    <option name="EXE_PATH" value="$PROJECT_DIR$/Samples/OneWayNotifications/bin/Debug/net7.0/OneWayNotifications.exe" />
+  <configuration default="false" name="Samples" type="DotNetProject" factoryName=".NET Project">
+    <option name="EXE_PATH" value="$PROJECT_DIR$/Samples/Samples/bin/Debug/net7.0/Samples.exe" />
     <option name="PROGRAM_PARAMETERS" value="" />
-    <option name="WORKING_DIRECTORY" value="$PROJECT_DIR$/Samples/OneWayNotifications/bin/Debug/net7.0" />
+    <option name="WORKING_DIRECTORY" value="$PROJECT_DIR$/Samples/Samples/bin/Debug/net7.0" />
     <option name="PASS_PARENT_ENVS" value="1" />
     <option name="USE_EXTERNAL_CONSOLE" value="1" />
     <option name="USE_MONO" value="0" />
     <option name="RUNTIME_ARGUMENTS" value="" />
-    <option name="PROJECT_PATH" value="$PROJECT_DIR$/Samples/OneWayNotifications/OneWayNotifications.csproj" />
+    <option name="PROJECT_PATH" value="$PROJECT_DIR$/Samples/Samples/Samples.csproj" />
     <option name="PROJECT_EXE_PATH_TRACKING" value="1" />
     <option name="PROJECT_ARGUMENTS_TRACKING" value="1" />
     <option name="PROJECT_WORKING_DIRECTORY_TRACKING" value="1" />
